Fix building on 64-bit Windows Review URL: http://codereview.chromium.org/546147
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@3700 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ test/cctest/test-debug.cc
@@ -2209,7 +2209,7 @@ bool GetEvaluateStringResult(char *message, char* buffer, int buffer_size) {
     return false;
   }
   Vector<char> buf(buffer, buffer_size);
-  int len = pos2 - pos1;
+  int len = static_cast<int>(pos2 - pos1);
   if (len > buffer_size - 1) {
     len = buffer_size - 1;
   }
